Refactor frame related items into a frame module
Move frame control logic from of rqd_servant to manager
