+++ config/rqd.fake_linux.yaml
@@ -10,6 +10,8 @@ machine:
   proc_stat_path: "/Users/dtavares/dev/rust_opencue/crates/rqd/resources/proc/stat"
   proc_loadavg_path: "/Users/dtavares/dev/rust_opencue/crates/rqd/resources/proc/loadavg"
   temp_path: "/tmp"
+runner:
+  snapshots_path: $HOME/.rqd/snapshots
 #   monitor_interval_seconds: 3
 #   use_ip_as_hostname: false
 #   nimby_mode: false
