+++ crates/rqd/src/servant/mod.rs
@@ -1,5 +1,5 @@
 /// Implement Rqd modules for rqd.proto's interfaces
-use crate::{config::config::Config, system::running_frame::RunningFrameCache};
+use crate::{config::config::Config, frame::manager::FrameManager};
 
 use opencue_proto::rqd::{
     rqd_interface_server::RqdInterfaceServer, running_frame_server::RunningFrameServer,
@@ -19,14 +19,14 @@ pub type Result<T> = core::result::Result<T, tonic::Status>;
 
 pub async fn serve(
     config: Config,
-    running_frame_cache: Arc<RunningFrameCache>,
     machine: Arc<MachineImpl>,
+    frame_manager: Arc<FrameManager>,
 ) -> Result<()> {
     let address: SocketAddr =
         SocketAddr::new(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0)), config.grpc.rqd_port);
 
-    let running_frame_servant = RunningFrameServant::init(Arc::clone(&running_frame_cache));
-    let rqd_servant = RqdServant::init(config, Arc::clone(&running_frame_cache), machine);
+    let running_frame_servant = RunningFrameServant::init(Arc::clone(&frame_manager));
+    let rqd_servant = RqdServant::init(config, machine, Arc::clone(&frame_manager));
 
     Server::builder()
         .add_service(RunningFrameServer::new(running_frame_servant))
