+++ crates/rqd/src/servant/rqd_servant.rs
@@ -1,17 +1,11 @@
 use std::fs;
-use std::panic::catch_unwind;
 use std::path::Path;
-use std::sync::{Arc, Mutex};
+use std::sync::Arc;
 
 use crate::config::config::Config;
+use crate::frame::manager::FrameManager;
 use crate::servant::Result;
-use crate::system::logging::FrameFileLogger;
 use crate::system::machine::Machine;
-use crate::system::running_frame::{RunningFrame, RunningFrameCache};
-use opencue_proto::host::HardwareState;
-use opencue_proto::job::nested_job::UidOptional;
-use opencue_proto::rqd::RunFrame;
-use opencue_proto::rqd::run_frame;
 use opencue_proto::rqd::{
     RqdStaticGetRunFrameRequest, RqdStaticGetRunFrameResponse,
     RqdStaticGetRunningFrameStatusRequest, RqdStaticGetRunningFrameStatusResponse,
@@ -28,30 +22,29 @@ use opencue_proto::rqd::{
     RqdStaticUnlockResponse, rqd_interface_server::RqdInterface,
 };
 use tonic::{Request, Response, async_trait};
-use tracing::error;
 
 pub type MachineImpl = dyn Machine + Sync + Send;
 
 /// Servant for the grpc Rqd interface
 pub struct RqdServant {
     config: Config,
-    running_frame_cache: Arc<RunningFrameCache>,
     machine: Arc<MachineImpl>,
+    frame_manager: Arc<FrameManager>,
 }
 
 impl RqdServant {
     pub fn init(
         config: Config,
-        running_frame_cache: Arc<RunningFrameCache>,
         machine: Arc<MachineImpl>,
+        frame_manager: Arc<FrameManager>,
     ) -> Self {
         if let Err(err) = Self::setup(&config) {
             panic!("Servant cannot start. Setup failed: {}", err)
         };
         Self {
             config,
-            running_frame_cache,
             machine,
+            frame_manager,
         }
     }
 
@@ -61,65 +54,6 @@ impl RqdServant {
         if !snapshots_path.exists() {
             fs::create_dir_all(snapshots_path)?;
         }
-        Ok(())
-    }
-
-    fn validate_frame(&self, run_frame: &RunFrame) -> Result<()> {
-        // Frame is already running
-        if self.running_frame_cache.contains(&run_frame.frame_id()) {
-            Err(tonic::Status::already_exists(format!(
-                "Not lauching, frame is already running on this host {}",
-                run_frame.frame_id()
-            )))?
-        }
-        // Trying to run as root
-        if run_frame
-            .uid_optional
-            .clone()
-            .map(|o| match o {
-                run_frame::UidOptional::Uid(v) => v,
-            })
-            .unwrap_or(1)
-            <= 0
-        {
-            Err(tonic::Status::invalid_argument(format!(
-                "Not launching, will not run frame as uid = {:?}",
-                run_frame.uid_optional
-            )))?
-        }
-        // Invalid number of cores
-        if run_frame.num_cores <= 0 {
-            Err(tonic::Status::invalid_argument(
-                "Not launching, num_cores must be positive",
-            ))?
-        }
-        // Fractional cpu cores are not allowed
-        if run_frame.num_cores % 100 != 0 {
-            Err(tonic::Status::invalid_argument(
-                "Not launching, num_cores must be multiple of 100 (Fractional Cores are not allowed)",
-            ))?
-        }
-
-        Ok(())
-    }
-
-    async fn validate_machine_state(&self, ignore_nimby: bool) -> Result<()> {
-        // Hardware state is not UP
-        if self
-            .machine
-            .hardware_state()
-            .await
-            .unwrap_or(HardwareState::Down)
-            != HardwareState::Up
-        {
-            Err(tonic::Status::failed_precondition(
-                "Not launching, host HardwareState is not Up",
-            ))?
-        }
-        // Nimby locked
-        if self.machine.nimby_locked().await && !ignore_nimby {}
-        // TODO: Chech if nimby is active and user activity was detected
-
         Ok(())
     }
 }
@@ -158,96 +92,7 @@ impl RqdInterface for RqdServant {
             .run_frame
             .ok_or_else(|| tonic::Status::invalid_argument("Missing run_frame in request"))?;
 
-        // Validate machine state
-        self.validate_frame(&run_frame)?;
-        self.validate_machine_state(run_frame.ignore_nimby).await?;
-
-        // Cuebot unfortunatelly uses a hardcoded frame environment variable to signal if
-        // a frame is hyperthreaded. Rqd should only reserve cores if a frame is hyperthreaded.
-        let hyperthreaded = run_frame
-            .environment
-            .get("CUE_THREADABLE")
-            .map_or(false, |v| v == "1");
-        let cpu_list = match hyperthreaded {
-            true => {
-                // On cuebot, num_cores are multiplied by 100 to account for fractional reservations
-                // rqd doesn't follow the same concept
-                let cpu_request = run_frame.num_cores as u32 / 100;
-                Some(
-                    self.machine
-                        .reserve_cpus(cpu_request)
-                        .await
-                        .map_err(|err| {
-                            tonic::Status::aborted(format!(
-                                "Not launching, failed to reserve cpu resources {:?}",
-                                err
-                            ))
-                        })?,
-                )
-            }
-            false => None,
-        };
-
-        // Although num_gpus is not required on a frame, the field is not optional on the proto
-        // layer. =0 means None, !=0 means Some
-        let gpu_list = match run_frame.num_gpus {
-            0 => None,
-            _ => Some(
-                self.machine
-                    .reserve_gpus(run_frame.num_gpus as u32)
-                    .await
-                    .map_err(|err| {
-                        tonic::Status::aborted(format!(
-                            "Not launching, insufficient resources {:?}",
-                            err
-                        ))
-                    })?,
-            ),
-        };
-
-        // Create user if required. uid and gid ranges have already been verified
-        let uid = match run_frame.uid_optional.as_ref().map(|o| match o {
-            run_frame::UidOptional::Uid(v) => *v as u32,
-        }) {
-            Some(uid) => self
-                .machine
-                .create_user_if_unexisting(&run_frame.user_name, uid, run_frame.gid as u32)
-                .await
-                .map_err(|err| {
-                    tonic::Status::aborted(format!(
-                        "Not launching, user {}({}:{}) could not be created. {:?}",
-                        run_frame.user_name, uid, run_frame.gid, err
-                    ))
-                })?,
-            None => self.config.runner.default_uid,
-        };
-
-        let running_frame = Arc::new(RunningFrame::init(
-            run_frame,
-            uid,
-            self.config.runner.clone(),
-            cpu_list,
-            gpu_list,
-            self.machine.get_host_name().await,
-        ));
-
-        self.running_frame_cache
-            .insert_running_frame(Arc::clone(&running_frame));
-        let running_frame_ref = Arc::clone(&running_frame);
-        // Fire and forget
-        let thread_handle = std::thread::spawn(move || {
-            let result = std::panic::catch_unwind(|| running_frame.run());
-            if let Err(panic_info) = result {
-                running_frame.update_exit_code(1);
-                error!("Run thread panicked: {:?}", panic_info);
-            }
-        });
-        // Another option would be to use a blocking context from tokio.
-        // let _t = tokio::task::spawn_blocking(move || running_frame.run());
-
-        // Store the thread handle for bookeeping
-        // TODO: Implement logic that checks if the thread is alive during monitoring
-        running_frame_ref.update_launch_thread_handle(thread_handle);
+        self.frame_manager.spawn(run_frame).await?;
 
         // Don't wait for frame to complete to return a response
         Ok(Response::new(RqdStaticLaunchFrameResponse {}))
