+++ crates/rqd/src/system/mod.rs
@@ -1,5 +1,2 @@
-pub mod frame_cmd;
 pub mod linux;
-pub mod logging;
 pub mod machine;
-pub mod running_frame;
